requirements.txt
openai
tiktoken
chromadb
langchain
python-dotenv
GitPython
streamlit